Refactor payment integration: replace Stripe service with new Payment service; add payment API handlers and models for checkout sessions and subscriptions
