+++ internal/payment/handler.go
@@ -1,4 +1,4 @@
-package stripe
+package payment
 
 import (
 	"figenn/internal/users"
@@ -20,7 +20,7 @@ func NewAPI(secret string, service *Service) *API {
 }
 
 func (a *API) Bind(rg *echo.Group) {
-	stripeGroup := rg.Group("/stripe")
+	stripeGroup := rg.Group("/payment")
 	stripeGroup.POST("/create-checkout-session", a.HandleCreateCheckoutSession, users.CookieAuthMiddleware(a.JWTSecret))
 	stripeGroup.GET("/subscriptions/:id", a.HandleGetSubscription)
 	stripeGroup.DELETE("/subscriptions/:id", a.HandleCancelSubscription)
